+++ History.md
@@ -1,3 +1,8 @@
+undefined
+=========
+
+  * add option "decimalPlaces"
+
 2.1.0 / 2015-05-21
 ==================